Disable drift detection for certain units
This commit is to disable drift detction for certain HR as otherwise due to various reasons the drift was being observed for  different components causing the deployment to fail.
The analysis of the reason and the fixes for each unit will be done in the separate issues and MR respectively.